Backports Issue 1117
Renaming variable "i" in the loop to avoid conflict with another "i" in
a different method. This should really be a local "var" variable, but
I'm not in a good position to test this, so, instead, making a
conservative change which will fix a corner case bug and won't break
others.