Making sure that BUILD_NUMBER is numeric
If A BUILD_NUMBER variable exists, it will be used as part of the
build directory name. It will also be included in the manifest files.
However, if BUILD_NUMBER is not numeric (e.g BUILD_NUMBER=2015-09-25)
the resulting json file will be invalid. This change aborts the
build if BUILD_NUMBER contains non-digit characters