Set the File umask in the Ruby environment at the start of execution so that all files created by Ruby or subcommands is created in a secure manner.
Some files like staging configuration and logs may not be created in a secure manner because the parent umask is not secure enough and we don't have enough information to know if it should locked down or not. In these cases, a warning message is displayed.

CONT-374

(cherry picked from commit a517975b1935ab9b3e4bc8a30e80b95366523a76)
Signed-off-by: Jeff Mace <jmace@vmware.com>