Updates CONT-1362
Corrected replicator so that tpm correctly sets the location of the
start-console.sh script and causes the replicator to fail if the script
cannot be found. This allows the replicator to mark obsolete plogs back
to the MINE process.

(cherry picked from commit 3931184314ba78af58180579e4820b2e647b94d6)
Signed-off-by: Jeff Mace <jmace@vmware.com>