CONT-1319
Made clear distinction between AuthenticationInfo and SecurityHelper so that AI includes all what is configured - which may include protocols and/or cipher suites which are not supported by SSL and thus they are not usable.
SecurityHelper always gives protocols and ciphers which can be used; they are never empty if encryption is used.

This is intermediate commit; getting supported protocols still doesn't work because SSLSocket is needed , and for that SSLSocketFactory must be created. SSLSocketFactory isn't created correctly because it would need key alias which isn't known.

(cherry picked from commit dc7179ca4cd23207ad5dc74098fc8a4b950679ac)
Signed-off-by: Jeff Mace <jmace@vmware.com>